doc: update job script section in flux-environment(7)
Problem: The documentation of the job script environment in
flux-environment(7) documents that `FLUX_JOB_RANKS` is only available
in the prolog and epilog, but it is now exported to housekeeping
scripts as well.

Update the docs to note that `FLUX_JOB_RANKS` is available in all
three script environments: prolog, epilog, and housekeeping.