broker: send SIGHUP to process group at broker exit
Problem: If a background process is executed from the broker or
an initial program such as the script of a batch job, then that
process may unexpectedly hold up job completion after the broker
exits because the job shell will be waiting for the process to close
stdio descriptors.

Since rc2 is now run in the broker's process group by default, attempt
to send SIGHUP from the broker to its own process group just before
exit. To avoid the broker being terminated by this SIGHUP, install
a one-time signal handler just before sending the signal.

Fixes #6725