testsuite: prepare #5105 reproducer for signal changes
Problem: The issue #5150 reproducer in the testsuite expects that
the initial program will exit with success (zero exit status) after
the job is signaled, but upcoming changes will allow signals to be
delivered to the initial program.

Do not depend on the exit status of `flux job status` in
t/issues/t5105-signal-propagation.sh.