t: fix potential race in backgrounded processes
Problem: The following pattern exists in a number of tests

do-something &&
do-something &&
run_a_process_in_background &
pid=$! &&
...
wait $pid / kill $pid / etc.

This is racy and unsafe.  In bash $! returns the pid of the most
recent backgrounded process, but with the previous line being && chained,
it is possible the backgrounding of the process has not yet occurred and
an incorrect PID is picked up by $!.

Solution: Remove the && chain for any command before a backgrounded
process.  In one situation, remove the workaround put in place to work
around this.

Fixes #6745