libeventlog: make eventlog formatter tests more reliable
Problem: The eventlog formatter unit tests compare raw (JSON) eventlog
output to static strings embedded in the test, but this can cause
spurious failures due to rounding or truncation of floating-point
representation in the generated JSON strings.

Convert the eventlog entry to JSON using json_dumps() before the
comparison to account for any idiosyncrasies of the current jansson
implemenation or platform.