libeventlog: fix uncalled error callback
Problem: The error callback in libeventlog's eventlogger is not
called when there is an error.  The issue arises from the fact
that the "current" batch is reset after a call to eventlogger_commit().
If an error occurs, eventlog_batch_get() is called to get the
"current" batch, which was just reset and is empty.  So no error callback
is issued.

Create a new wrapper function commit_batch() that is identical to
eventlogger_commit(), but also returns the value of the recently
committed batch.  Use this value instead of calling eventlog_batch_get().

Fixes #6753