task-exit: improve clarity of log message
Problem: it has been requested that we kindly excise the word "doom"
from the exit timeout message because it is "EXTREMELY confusing" [sic].

OK. Rename the doom shell plugin to task-exit so that the log message
prefix isn't so distracting.

In case the message was actually unclear for other reasons, add the
word "task" before "rank" to make sure we communicate that a task
exited and not a broker, shell, or other ranked entity.

Fixes #6749