python: modify flux-fortune argparse usage
Problem: with the addition of an argparse epilog for CLI plugins
help output, the flux-fortune command now throws an error since it
does not have the ability to take an epilog (or an epilog).

Allow arbitrary keyword args to be passed down to the flux-fortune
command's argparse parser. This will not cause the output to change.