testsuite: test conent.restore=auto with no RESTORE link
Problem: None of the tests in the testsuite test starting an instance
with content.restore=auto but no RESTORE link, but this is the case
when first starting a system instance, and exercises key login in
etc/rc1.py.

Add -Scontent.restore=auto to the first instance run in
t2810-kvs-garbage-collect.t so that this case is exercised in the
testsuite.