kvs: remove unnecessary kvsroot lookup
Problem: When the kvs module is initially loaded, the primary KVS
namespace root is looked up before it is created.  There is no point
to this lookup, since we are initializing the KVS.  This appears to
be a cut and paste from other areas of the KVS where a lookup of a
kvsroot would make sense before creating it.

Solution: Remove this unnecessary kvs root lookup.