python: add jobspec `__repr__` method
Problem: when debugging or writing Flux scripts with the
Python API, it's often useful to look at the JSON/dictionary
representation of a jobspec and inspect it. Normally, this
requires remembering the top-level keys and accessing them
directly, or using `dir()`.

Add a `__repr__` method to Jobspec to improve our lives.