flux-fsck: add basic fsck to check content data
Problem: There is no basic tool to verify if content data is
valid or corrupt.

Add a basic flux-fsck tool that will walk content data, starting
with it's checkpoint, and ensure that all references are valid
and retrievable.  Output any errors that are found along the way.

This initial fsck tool is very basic and is considerd a starting point
for more advanced features down the road.