flux-fsck: add basic fsck to check KVS backing
Problem: There is no basic tool to check the integrity of
the data in the KVS backing store.

Add a basic flux-fsck tool that will walk content data, starting
with the most recent checkpoint, and ensure that all references are valid
and retrievable.  Output any errors that are found along the way.

This initial fsck tool is very basic and is considerd a starting point
for more advanced features down the road.