job-exec: add `kill-shell-delay` configuration option
Problem: The delay that the job execution system uses to transition
from killing the job tasks to the job shell is hardcoded to
5*kill-timeout and can not be configured.

Add a `kill-shell-delay` configuration option and module parameter
that can optionally be used to set the delay from the time the first
term signal is sent to when the job-exec module transitions to killing
the job shell. The default is 5 * kill-timeout.

Expose the currently configured `kill-shell-delay` in the job-exec
module stats. For jobs, also expose the remaining time until the kill
shell timer elapses.