job-manager: avoid processing jobs multiple times on restart
Problem: On restart, the job manager iterates over active jobs to
process jobtap dependencies, the job.new callback, and also to post
the flux-restart event to jobs in the SCHED state. The iteration is
performed using `zhashx_first()` and `zhashx_next()`, but this is not
safe when the hash is modified, which must be happening when there
are job dependencies, because jobs were getting multiple flux-restart
events posted to the job eventlog. Debugging showed that the same
jobs were being processed in this loop multiple times.

Grab a list of active jobs using `zhashx_values()` and use that for
iteration instead of iterating the hash directly.