jobtap: avoid posting duplicate dependency-add events
Problem: The `job.dependency.*` callbacks for notifying jobtap
dependency plugins of job dependencies are called not only when a new
job is submitted, but also during job manager restart. In the restart
case, a `dependency-add` event is likely already in the job eventlog
and registered with the job manager since all events in the KVS have
been processed before initializing jobtap pluging, but a dependency
plugin can't determine the difference between normal submission and
restart, so it will likely re-issue the dependency-add event.

To avoid duplicate `dependency-add` events in job eventlogs after
restart, avoid actually posting the event when the dependency is
already registered in the job's dependency set.