job-manager: dequeue jobs from stopped queues on restart
Problem: The job manager doesn't dequeue jobs from queues stopped at
restart. This allows jobs in the anonmymous queue, which is created
in the "started" state, to run on restart even when the queue is
stopped before shutdown.

Move the `dequeue_jobs()` call into `queue_stop_all()` and add a new
`queue_stop_one()` that does the same but for the single named queue
being stopped. This should prevent omission of this call on future
uses of these functions. Update callers accordingly.

Fixes #6789