broker: add 'mincrit' topology
Problem: kary:0 usefully minimizes critical nodes for batch jobs but may
not perform well.

Add a new TBON topology 'mincrit[:K]' for defining a 3 level tree where the
fanout from leader to routers is K, and the fanout from routers to leaves
is dependent on the size.

If K is unspecified, a value is chosen that keeps the fanout to the leaves
at or below 1024.  If there are 1025 nodes or less, a flat topology is used.

Example:

  $ flux start -o,-Stbon.topo=mincrit:2 --test-size=11 flux overlay status
  0 system76-pc: full
  ├─ 1 system76-pc: full
  │  ├─ 3 system76-pc: full
  │  ├─ 5 system76-pc: full
  │  ├─ 7 system76-pc: full
  │  └─ 9 system76-pc: full
  └─ 2 system76-pc: full
     ├─ 4 system76-pc: full
     ├─ 6 system76-pc: full
     ├─ 8 system76-pc: full
     └─ 10 system76-pc: full

Fixes #6807