shell: allow taskmaps to redistribute tasks amongst nodes
Problem: taskmap_check() now allows an updated taskmap to change
the counts of tasks per node, if total task and node count remains
unchanged, but the affected rcalc data in the shell is not updated,
leading to a possible mismatch between various shell components.

Add a rcalc_update_map() function which can be passed a new taskmap
and updates ntasks on each shell rank internally.

Also update the rankinfo object for the current shell after
rcalc_update_map() returns.

This should now allow a custom taskmap to redistribute task counts
amongst the nodes of a job.