testsuite: update shell taskmap tests
Problem: No tests in the testsuite check that a taskmap which
changes task counts amongst nodes works with the job shell. Also,
one expected failure now succeeds because this type of taskmap is
now allowed.

Remove the failing test.

Add a test that assigns 3 tasks to rank 0 and one task on other
nodes.