testsuite: workaround possible bind() to hostname interface
Problem: A test in t0019-tbon-config.t fails in github ci with an error

 error binding to tcp://10.1.1.7:*: Cannot assign requested address

This occurs when `--network=host` is used with docker in the github
runner environment for unknown reason. Work around the failure by
allowing the test to pass either if the `flux start` invocation
succeeds, or if the test encounters the error above.

Fixes #6812