flux-kvs: support sync command
Problem: The only way to perform a sync is through the
`flux kvs put` command, which inherently requires the user to
also write data at the same time.  This isn't necessarily always
desired.

Support a new `flux kvs sync` command.  It is largely identical
to `flux kvs put --sync`, but does not require additional data
to be written.

Fixes #6814