doc: add note on use of flux queue drain
Problem: It is common to use `flux queue drain` to wait for all
submitted jobs to finish in flux-batch(1) scripts.  However, this
may not be obvious to new flux users.

Add a note about this in flux-queue(1) if users look for information
on why this is used in flux batch scripts.