broker: use bizcard in overlay code
Problem: the PMI bootstrap and overlay code for the broker's
public key, hostname, and bind URI is rather messy and is going
get worse if support for mutiple bind URIs is added in the same
style.

Store a bizcard container for the local broker in the overlay context
instead of a la carte data.  This object can be shared directly with
the PMI bootstrap code.  It also prepares the overlay code for
binding to multiple endpoints.