python: add accessor for BatchConfig cli class
Problem: flux.cli.base.BatchConfig is used as a convenience class
for managing a Flux config object built up from multiple uses of the
`--conf` option with `flux batch` and `flux alloc`, but it lacks a
convenient accessor for the currently set conf keys.

Add a get() method to the class which tasks a dotted key and returns
the current config at that key or a default value. This will make it
simple for callers to get any current config value before performing
further updates.