broker: use both IPC and TCP within an instance
Problem: when brokers are co-located and bootstrap with PMI,
they currently only use ipc if the entire instance is on one
node, but instances where some brokers are co-located could
benefit from IPC too.

Use the taskmap to determine when peers are co-located and
use IPC between those, unless it is suppressed by setting the
tbon.prefertcp broker attribute to a nonzero value.