cmd: support addition of node 0 brokers in flux-batch/alloc
Problem: For subinstances of Flux with large node counts, it may be
useful to launch an extra set of brokers on node 0 of the allocation
to assist with handling message distribution, but there's no easy
way for users to request that in flux-alloc and flux-batch.

Add a new `--add-brokers=N` option to `flux alloc` and `flux batch`
which requests that `N` extra brokers be started on node 0 of the
allocation.  The option is hidden and undocumented for now in case
a better solution is implemented in the future. The `--add-brokers`
option is only available when a number of nodes is explicitly requested
with `-N, --nodes`.

The `--add-brokers` option adjusts the job's taskmap to force extra
brokers onto the first allocated node and automatically excludes the
extra ranks to ensure their resources are not available for scheduling.
The jobspec is also updated to reflect the updated task count.