broker: use taskmap to determine best iface
Problem: it's hard to write single node CI tests for the ability
to use both ipc:// and tcp:// endpoints when using business card
hostnames to determine locality.

Refactor some internal functions to make the taskmap object available
in the main line PMI code, and then use that to determine locality.