job-manager: add dependency-singleton jobtap plugin
Problem: There's no support for a singleton job dependency scheme,
i.e. one in which only one job of a given name is allowed to be
running at time for the same userid.

Add a dependency-singleton jobtap plugin which supports a singleton
dependency scheme. This plugin tracks the count of all active jobs
with an explicit job name per userid. Jobs submitted with the singleton
dependency scheme are then placed in a list and held in the DEPEND
state until other jobs with the same name and userid (and not already
held in the singleton list) become inactive.

Fixes #6804