libjob: add struct for non-integer counts
Problem: to expand implementation of RFC 14, several components in core
will need to handle non-integer counts, but there is currently no support.

Add a new `count` wrapper struct to libjob that will contain one of an
integer, RFC 14 range, or RFC 22 idset, with an interface modelled as a
subset of the libidset interface. This commit implements the initial
decoding and encoding infrastructure.