configure: check for Lua posix module
Problem: The flux-core configure will pass all checks when the Lua
posix module is not available, but subsquently `make check` will fail.
While strictly not a build requirement, it would reduce user confusion
if configure ensured presence of this required Lua module.

Add a check for the Lua posix module in configure.