README.md: Add tips for `make check`
Problem: `make check` in flux-core takes a very long time to complete
without use of `-j`, but the README doesn't mention or suggest use
of this option.

Add `-jN` to the options of both `make` and `make check` in the
instructions. Also add a tip suggesting the use of `-j` with `make
check` specifically, and add a link to the testsuite README.