content-sqlite: store history of checkpoints
Problem: The content-sqlite backend only stores one checkpoint,
overwriting the older checkpoint everytime a new checkpoint is
written.  If anything goes wrong with a checkpoint, there is no
information stored to allow a recovery from an earlier point in time.

Instead of overwriting checkpoints, store a rolling history of the
checkpoints.  Support through a new "checkpt_v2" table.

Support configuration of the maximum number of checkpoints via a new
configuration parameter.  The default maximum number of checkpoints
is 5.

Fixes #6629