testsuite: fix potential failure in t2602-job-shell.t
Problem: On some systems, the test `job-shell: fails on modified J`
fails in t2602-job-shell.t with the error:

 flux: Argument list too long

because the modified `J.new` is too long for shell argument processing.

Provided the `J.new` content on stdin of `flux kvs put` instead of
trying to pass it as an argument.