python: add extra parameters to JobspecV1 factory methods
Problem: Common jobspec attributes and options are not available
in the JobspecV1 factory methods, `per_resource`, `from_command`,
`from_batch_command` and `from_nest_command`. This unnecessarily
requires the user to set some critical attributes of the resulting
jobspec after it is returned, for example the working directory and
environment, which makes the interface awkward and error prone.

Add a set of the likely most commonly used extra parameters to
`JobspecV1.from_command()`. Pass these args along in `**kwargs` from
the other factory methods so the extra parameters are supported there
as well.

Update `from_command()` docstring to the Google style and document
the new args. Document `**kwargs` in other functions and reference
the docs in `from_command()`.