t: remove key from checkpoint tests
Problem: The key in content checkpoint put/get is now optional,
so it is no longer needed in the majority of content checkpoint
tests.

Update functions in content-helper.sh to not take a key.
Update tests in t0012-content-sqlite.t and t0018-content-files.t
to no longer specify a key.
Add legacy functions that support a key for legacy "key" tests.
Remove one now irrelevant test in t0012-content-sqlite.t.