testsuite: reduce parallelism in t4000-issues-test-driver.t
Problem: The issues tests in t4000-issues-test-driver.t are run as
jobs in a test instance of size=2, but the instance inherits the
total number of cores from the system, which means up to 2*ncores
issues tests could be run at the same time. On some systems, this
causes failures in the tests due to overloading the system.

Switch to the "job" personality of test_under_flux, which will
configure the test instance with only 2 fake cores per rank, keeping
the parallelism of the tests to 4, which is probably a safe number.

Fixes #6856