t: fix race in job info update watch test
Problem: In a test in t2233-job-info-update.t there is a small
race possibility.  An update watcher assumes that an update
has "propagated" and that an earlier value of that update cannot
be seen.  However, a recent test failure indicates that the earlier
value can be seen.

Solution: Move one of the "waits" earlier in the test to ensure
the updated value has fully "propagated".  This ensures later
watchers will only get the updated value.

Fixes #6838