delegate: allow job delegation to a different flux instance
Problem: Delegating a submitted job to a different flux
instance is currently not supported. This feature is useful
for users that have jobs that need to be executed in a
multi-cluster setup, or for workflows where a certain
sub-job needs to be executed on a peer cluster.

Allow delegation of jobs to a different, peer-level flux
instance by utilizing the URI of the other instance.
Currently, explicit checking for feasibility of the job
on the delegated instance is not performed, and it is
assumed that both the instances belong to the same user.