python: unify behavior of Jobspec `getattr` and `setattr`
Problem: The Jobspec getattr() and setattr() methods to get and set
jobspec attributes do not behave similarly. While both methods default
to an `attributes.` prefix if not provided, setattr() further defaults
to an `attributes.system.` prefix if the prefix is not `user.` or
`system.`, while getattr() does not.

Unify the behavior of these two function so they both default to
`attributes.system.`. Abstract the prepend of the default key path
to a helper function to reduce duplicate code.