python: rename Jobspec stdio properties
Problem: The Jobspec stdio property getters and setters have
names that do not match the CLI options, which may be not be ideal.

Rename the Jobspec stdio properties from `stdin`, `stdout`, `stderr`,
to `input`, `output`, and `error` to match the respective cli options.
Keep the existing properties as aliases for the new names.