kvs-watch: support initial sentinel option
Problem: When watching appends to a key, it would be convenient if
we knew when the "initial" or "current" data has been fully sent before
waiting / watching for future data.  It would take several RPCs to
accomplish this at the moment.

Support a new KVS_WATCH_INITIAL_SENTINEL flag in kvs-watch.
An empty RPC response will be sent after all "initial" or "current"
data has read from the KVS and sent to the caller.

Update flux_kvs_lookup_get() and flux_kvs_lookup_get_raw() to return
NULL (or 0) for OUT parameters when a sentinel has been received.