broker: let a upmi client relay extra bootstrap info
Problem:  there is no way for a upmi client to relay auxiliary bootstrap
information through the broker to applications.

Add an optional dict plugin OUT argument that can be set by the
upmi_initialize() implementation.  A new dict member is added to
struct pmi_info that is set to a json dictionary when the implementation
sets  the OUT argument.  In the broker, any key-value pairs in the dict are
directly set as broker attributes.