cmd: add flux-multi-prog.py
Problem: Flux does not currently support for MPMD (multiple program
multiple data) execution for parallel jobs,

Add a new command `flux multi-prog` which takes a mandatory config
file which maps task rank to program and arguments. This utility
is then used as the command and arguments to `flux submit` and/or
`flux run`, i.e.

 $ flux run -N4 flux multi-prog test.conf