resource: log drain/undrain events
Problem: It may be useful at some sites to capture drain and undrain
events in syslog for auditing and search, but currently these events
go only to the resource eventlog.

Log drain and undrain events in the resource module just after they
are posted to the eventlog so that they can be captured by syslog.

Fixes #6878