broker: set jobid-path attribute
Problem: there is no unique identifier for jobs in a Flux instance
hierarchy.

As a preliminary step, if FLUX_JOB_ID_PATH is set in the broker's
environment, set the jobid-path broker attribute to its value
and unset the environment variable so that it is not inadvertantly
inherited by the initial program (like how FLUX_JOB_ID is handled).

If FLUX_JOB_ID_PATH is not set, set the broker attribute to "/".