libutil: fix potential hang in cronodate_next()
Problem: When incrementing the weekday member of a `struct tm`, the
`tm_mday` value was assigned instead of incremented, which could make
the `struct tm` go back in time instead of forward as cronodate_next()
assumes. With certain initial conditions, this can cause the loop in
cronodate_next() to bounce between values and never terminate.

Replace the assignment with an increment to fix the issue.

Fixes ##6886