broker: add broker.getenv RPC
Problem: in nested Flux instances, a shell plugin may need
to access environment variable settings for the local broker.

Specifically, a cray-slingshot plugin needs to find inherited
CXI services, which may differ on each broker.

Add a new broker RPC: broker.getenv.  A user sends an array
of key names and gets back a dict of key/value pairs.  Any
environment variables for which getenv(3) returns NULL are
omitted from the object.

The RPC is restricted to the instance owner.