rc: require opt-in to use flux-modprobe(1)
Problem: The new modprobe based rc scripts break other framework
projects, which need time to transition to the new rc init scheme.

Require an opt-in to using flux-modprobe(1) in rc1 and rc3 by setting
the environment variable FLUX_RC_USE_MODPROBE.

Set this variable in t/sharness.d/flux-sharness.sh so that modprobe
is used in the testsuite.

Similarly set the variable in fluxometer.lua and t/python/subflux.py.