broker: associate module argv with module object
Problem: in the future, module thread cleanup will need
to occur in a cancellation callback, but the module
argv is allocated in the module thread and would be leaked
in such an arrangement.

Add argc and argv as module struct members and alloc/free
them in the broker before the module thread starts.
There was really no need to do that in the module thread.