broker: associate module errno with module object
Problem: in the future, module thread cleanup will need
to occur in a cancellation callback, but the module
errno (if any) is on the stack of the module thread
and would not be available in a cleanup callback.

Add mod_main_errno and mod_main_failed as module struct
members so they can be easily passed to the cancellation
callback.