broker: add a timeout to broker notification
Problem: a module expects a response to its request to
change the module state to FINALIZING, but there won't be
one if the request is made in a cancellation cleanup callback
and the cancellation occurs after the broker reactor has
exited.

Wait at most 1s for this response, then move on.