broker: support module thread cancellation
Problem: when a broker module is stuck looping somewhere other
than the reactor, flux module remove's reactive shutdown request
is ignored.

Support an optional cancel flag in the module.remove request.
When this flag is set, pthread_cancel(3) is called on the module
thread.  As long as the thread's loop is passing through one of
the cancellation points listed in pthreads(7), cancellation should
be effective.

Fixes #6887