jobtap: add `flux_jobtap_call()`
Problem: There's no way for a jobtap plugin to make a call into the
plugin stack, but this may be useful when plugins themselves which
to support extensions, or when a plugin is split into multiple parts.

Add `flux_jobtap_call()`, which simply offers the jobtap call
interface to plugins. Since much of the jobtap API assumes there is a
context of a current job, require a job id or `FLUX_JOBTAP_CURRENT_JOB`
as a `flux_jobtap_call()` argument.

Fixes #6895