testsuite: add tests for `flux_jobtap_call()`
Problem: There's no testing of the `flux_jobtap_call()` function.

Add a test jobtap plugin that invokes `flux_jobtap_call()` along with
a separate plugin that registers a callback for the called topic.
Load the test plugins in `t2212-job-manager-plugins.t` and run a job
to ensure the interface works.

Additionally add some invocations of `flux_jobtap_call()` with bad
arguments to the `jobtap_api.so` test jobtap plugin (already used in
the testsuite)