resource: avoid losing R expiration on resource.rediscover
Problem: When a job with a time limit is started with the config key
`resource.rediscover=true`, the resource module replaces the parent
R with an R discovered by hwloc. In doing so, however, any expiration
given the job is lost, so the scheduler and other components do think
the resources are granted without expiration. This leads to incorrect
default timelimits and incorrect results from `flux job timeleft`
and other time limit tools like libyogrt.

When resource.rediscover is set, save the expiration from R in the
inventory so it can be re-established before the rediscovered inventory
is placed in the KVS.

Fixes #6901