content-sqlite: add comment about no sqlite3 reset
Problem: In content_sqlite_load(), we do not reset the load stmt.
This seems strange at first, but it is necessary for the caller to
reset the statement because we return data from the function for
the caller that is dependent on the prior sqlite query.

Add comment in content_sqlite_load() to clarify the lack of calling
sqlite3_reset() on the load statement.