flux-fsck: use validate for valref blobref checks
Problem: When validating the references in a valref, the content.load
service is used.  This can be inefficient because content.load will return
all data from the content backing store, but this data isn't needed or used.
We are only trying to verify if the references is valid and the data
is backed up.

Use the new content-backing.validate services instead of a load.  Update
a number of function names from "load" to "validate" as a result.