docker: remove libmpich-dev from noble image
Problem: in Ubuntu Noble, the apt-installed MPICH is built such
that libmpich.so uses PMIx for bootstrapping, but Hydra is still used
as the MPI launcher. Since Hydra does not provide a PMIx interface,
this prevents bootstrapping of MPI processes from working correctly.

Solution: remove libmpich-dev from the list of apt packages. This forces
applications to build against the from-source build of MPICH, which
does not have the problem above. Also, add installs of gfortran
to prevent configure errors when building MPICH from source.