etc: only try to run regular executable files in rc1
Problem: rc1 and rc3 use the test `-e $rcfile` to determine if they
should execute a file when looping over files in `$rcdir/rc1.d/*` or
`$rcfile/rc3.d/*`. But this just checks if the file exists, which of
course it does because the file was returned via the glob. This results
in rc1/rc3 attempting to execute directories, which results in instance
failure. Note: This can now occur when Python modprobe scriptlets
are placed in these directories as supported by flux-modprobe(1).

Assume that the check here was meant to check that entries matching
the glob are a regular file and executable (`-f` and `-x`). Update some
tests in the testsuite that did not make the rc scripts executable.