python: handle string passed to flux.cli.base.list_split()
Problem: If a string instead of a list of strings is passed to
flux.cli.base.list_split(), then the string ends up being split into
characters instead of split on comma.

Protect against this usage by calling list_split([arg]) if the argument
is specifically an instance of `str`.