broker: refactor service registration
Problem: module send/recvs are not all in one place, making
adding a trace point slightly more complex than it needs to be.

Refactor so that all calls to module_sendmsg_new() and
module_recvmsg(), and module_event_cast() are in modhash.c.
This required the code for registering a service name to move
into modhash.c.