broker: match module names in trace request
Problem: the module trace code ignores the names[] array
in the trace request because it assumes the trace point is
within one module, thus must always match.  This assumption
will no longer be true when the trace point is moved to the
broker end.

If names[] is non-empty in the request, ensure that the module
name that is passed in at the trace point matches one of the
names[], otherwise suppress the trace.