broker: trace modules on the broker end
Problem: early messages of a broker module cannot be traced.
The module must be loaded before a trace request can be accepted
for it, since the tracing occurs within the module thread
and state is kept in the module_t struct which doesn't exist
before the module is loaded.

Handle module trace requests in the broker rather than the module
thread, and keep the state in modhash_t instead of module_t.
Now a trace request can be accepted before the module is loaded
and early messages can be captured.

Fixes #6770