libkvs: support multiple checkpoints in KVS API
Problem: The internal KVS checkpoint API only supports the retrieval
of a single checkpoint.  However, some content backing modules may
store multiple checkpoints.

Update the KVS checkpoint API to support an array of checkpoints
to be returned and parsing functions to parse the individual
entries in the array.  Note that the content backing modules do not
yet support returning multiple entries.  This update is in preparation
for that future change.

Update all callers accordingly.