testsuite: add debug to some sporadically failing tests
Problem: A few tests are failing as reported in issue #6908, but
there's no debug information or other output to help diagnose the
reason for the failures.

Add extra debug to specific failing tests in t2603-job-shell-initrc.t,
t2613-job-shell-batch.t, and t3002-pmi.t to assist debug of these
failures.