broker: avoid extra malloc in log append
Problem: the use of zlist_t in the logging ring buffer
implementation means that a list node has to be separately
allocated each time a log entry is appended to the ring.

Switch to a CCAN list which embeds the list node in the
log entry struct and thus avoid the extra malloc call.