ipaddr: use macos interface to get default if
problem: MacOS/BSD lack a linux-compatible proc (usually) so we can't
read routes from proc

solution: On linux we'll continue to use /proc/net/route, while on macos
it turns out you can have a default interface *that does not have a
netmask of zero*.  To get it reliably, I've added a version of
`getprimary_interface4` that uses the CoreFoundation and
SystemConfiguration frameworks to request the global IPv4 default
interface name from macos directly.  This is more macos specific than
using a PF_ROUTE socket, but much _much_ simpler.  If we want this to
work on other BSDs (aside from FreeBSD with linux compat turned on)
we'll want to do that but NETLINK/PF_ROUTE are really arcane and rather
high overhead for something this simple.

Anyway, this is using the classic C interface to CoreFoundation rather
than any of the add-ons so there's no need to compile code as
Objective-C or anything.

As to why I switched to this from the getifaddrs approach, that seemed
to work fine on linux with a small tweak, but not on macos, and if it
doesn't work reliably cross-platform it seems better to keep what we had
battle tested and add the more specific mac version like this.