etc: install modprobe rc1.d and rc3.d directories
Problem: flux-modprobe(1) supports Python extensions placed in
/etc/flux/modprobe/rc{1,3}.d and extra TOML config in modprobe.d,
but because flux-core does not provide any files in these directories
they are not managed by `make install` and `make uninstall`.

Create empty *.d directories under `$sysconfdir/flux/modprobe` on
installation and remove them (and empty modprobe dir) on uninstall
so that it is obvious these directories are available and managed
by flux-core.