broker: don't require broker modules to be a DSO
Problem: the module code assumes all broker modules are
loaded from DSOs.

Move the DSO code to module_dso.[ch] and just pass mod_main()
to module_create().  Put the dlopen() handle in the module's
aux container for destruction when the module is destroyed.

This prepares the way for builtin modules in a subsequent commit.