broker: define builtin modules
Problem: broker modules are required to be DSOs, even if
they are always loaded, which is inefficient and awkward.

Define a list of builtin modules in modhash.c and arrange
for them to be automatically loaded before rc1 executes and
unloaded after rc3 executes.

A builtin module is built as a convenience library that
defines a public symbol with type 'struct module_builtin'.
This struct includes the module name and a pointer to its
mod_main() function.

Convert connector-local to a builtin broker module.