shell/test: build test modules without version
problem: the plugins built specifically for the plugstack tests weren't
getting the avoid-version flag to libtool, so they were generating
symlinks to *.so.<ver> on linux and *.<ver>.so on macos.  On macos this
caused the plugstack tests to fail due to more libraries being found
than should be.

solution: add the avoid-version flag to the libraries in automake
(thanks for this one @grondo, I was really lost)