subprocess/local: avoid posix_spawn on macos
problem: As @garlick noted we get EBADF when trying to run local
processes through subprocess. I _think_ this is because the macos
implementation reports errors when asked to close a file descriptor that
is already closed.

solution: Because of the structure of the spawn handler, and the likely
low need for high-speed launch on macos, I decided to force macos to use
the fork launcher for now. This should be revisited, as should allowing
posix_spawn to chdir if we have support for that in the linux versions
we target.